refactor: separate login and sign up forms, move switch to login page instead of in the component
