+++ pages/login.vue
@@ -10,9 +10,26 @@ definePageMeta({
     },
   ],
 })
+
+const isSigningIn = ref(false)
+
+function handleSwitchForm() {
+  isSigningIn.value = !isSigningIn.value
+}
+
+const title = computed(() => (isSigningIn.value ? "Log in" : "Sign in"))
+useHead({ title })
+
+const switchFormButtonLabel = computed(() =>
+  isSigningIn.value ? "Belum punya akun? ayo bikin dulu!" : "Sudah punya akun? Masuk"
+)
 </script>
+
 <template>
   <div class="flex justify-center min-h-[70vh]">
-    <SignInForm />
+    <CTA type="button" @click="handleSwitchForm" :label="switchFormButtonLabel" />
+
+    <AuthSignInForm v-if="isSigningIn" />
+    <AuthLoginForm v-else />
   </div>
 </template>
